Integrate Vehicle Exit Check directly into Patch_FormCaravanDialog and TransferableVehicleWidget
