+++ 1.5/Languages/English/Keyed/Caravans.xml
@@ -10,7 +10,6 @@
 	<VF_ImpassableBiome>Biome impassable for this vehicle.</VF_ImpassableBiome>
 	<VF_CaravanDisabled>Caravaning Disabled.</VF_CaravanDisabled>
 	<VF_CaravanCantMove>Unable to move.</VF_CaravanCantMove>
-	<VF_CaravanCantReachExit>Unable to reach map edge.</VF_CaravanCantReachExit>
 	<VF_CantMergeAerialVehicle>Can't merge aerial vehicles.</VF_CantMergeAerialVehicle>
 	
 	<VF_CommandDockShip>Dock</VF_CommandDockShip>
